backend/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asyncer==0.0.8
attrs==25.3.0
autogen==0.8.5
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comtypes==1.4.10
dataclasses-json==0.6.7
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
elevenlabs==1.56.0
emoji==2.14.1
faiss-cpu==1.10.0
fastapi==0.115.12
filelock==3.18.0
frozenlist==1.5.0
fsspec==2025.3.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.1
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.51
langchain-openai==0.3.12
langchain-text-splitters==0.3.8
langsmith==0.3.24
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.3.2
mypy-extensions==1.0.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.4
openai==1.70.0
orjson==3.10.16
packaging==24.2
propcache==0.3.1
pyautogen==0.8.5
pydantic==2.11.2
pydantic_core==2.33.1
pydantic-settings==2.8.1
pygame==2.6.1
pypiwin32==223
python-dotenv==1.1.0
pyttsx3==2.98
pywin32==310
PyYAML==6.0.2
rake-nltk==1.0.6
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.5.3
setuptools==65.5.0
sniffio==1.3.1
SpeechRecognition==3.14.2
SQLAlchemy==2.0.40
starlette==0.46.1
sympy==1.13.1
tavily-python==0.5.4
tenacity==9.1.2
termcolor==3.0.1
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.50.3
typing_extensions==4.13.1
typing-inspect==0.9.0
typing-inspection==0.4.0
urllib3==2.3.0
uvicorn==0.34.0
websockets==15.0.1
yarl==1.18.3
zstandard==0.23.0
